% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotInx.R
\name{PlotCCInx}
\alias{PlotCCInx}
\title{Plot cell-cell interactions as bipartite graph}
\usage{
PlotCCInx(INX, cellTypeA, cellTypeB, proteinTypeA, proteinTypeB,
  GeneMagnitudeThreshold, GeneStatisticThreshold, TopEdges, GeneNames,
  YSpacing = "relative")
}
\description{
Plot cell-cell interactions as bipartite graph
}
