% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViewAgingMouseBrainCCInx.R
\name{ViewAgingMouseBrainCCInx}
\alias{ViewAgingMouseBrainCCInx}
\title{View AgingMouseBrain cell-cell interaction predictions in the CCInx Shiny app}
\usage{
ViewAgingMouseBrainCCInx()
}
\value{
The function causes the scClustViz Shiny GUI app to open in a
  seperate window.
}
\description{
A wrapper function to view the \code{AgingMouseBrainCCInx} dataset in the
\code{CCInx} Shiny app.
}
\examples{
  ViewAgingMouseBrainCCInx()

}
\seealso{
\url{https://baderlab.github.io/CCInx} for information on
  \code{scClustViz}.
}
